+++ rotkehlchen/chain/gnosis/transactions.py
@@ -35,6 +35,8 @@ class GnosisWithdrawalsQueryParameters:
     last_block_processed: int
     expected_topics: set[str]
     gnosis_transactions: 'GnosisTransactions'
+    addresses: list[ChecksumEvmAddress]
+    from_ts: Timestamp
 
 
 def _process_withdrawals_events_cb(
@@ -56,7 +58,27 @@ def _process_withdrawals_events_cb(
                     associated_address=bytes32hexstr_to_address(event['topics'][1]),
                     must_exist=True,
                 )
+    try:
+        block = cb_arguments.gnosis_transactions.evm_inquirer.get_block_by_number(
+            num=last_block_queried,
+        )
+        last_queried_ts = block['timestamp']
+    except (KeyError, RemoteError) as e:
+        msg = f'Missing key {e}' if isinstance(e, KeyError) else str(e)
+        log.error(f'Failed to query block timestamp for gnosis bridge for {cb_arguments.addresses} due to {msg}')   # noqa: E501
+        return
 
+    db = cb_arguments.gnosis_transactions.database
+    with db.user_write() as write_cursor:
+        for address in cb_arguments.addresses:
+            db.update_used_query_range(
+                write_cursor=write_cursor,
+                name=f'{BRIDGE_QUERIED_ADDRESS_PREFIX}{address}',
+                start_ts=cb_arguments.from_ts,
+                end_ts=last_queried_ts,
+            )
+
+    log.debug(f'Saved gnosis logs range from {cb_arguments.from_ts} to {last_queried_ts}')
     cb_arguments.last_block_processed = last_block_queried
 
 
@@ -103,9 +125,10 @@ class GnosisTransactions(EvmTransactions):
             last_block_processed=NO_BLOCK_PROCESSED_VALUE,
             expected_topics={'0x000000000000000000000000' + x.lower()[2:] for x in addresses},
             gnosis_transactions=self,
+            addresses=list(addresses),
+            from_ts=from_ts,
         )
 
-        last_queried_ts = None
         try:
             self.evm_inquirer.get_logs(
                 contract_address=BLOCKREWARDS_ADDRESS,
@@ -115,28 +138,9 @@ class GnosisTransactions(EvmTransactions):
                 argument_filters={'receiver': addresses[0]} if len(addresses) == 1 else {},
                 from_block=from_block,
                 to_block='latest',
-                call_order=self.evm_inquirer.default_call_order(),
+                call_order=None,  # Intentionally None here to use Etherscan and Owned nodes only
                 log_iteration_cb=_process_withdrawals_events_cb,
                 log_iteration_cb_arguments=callback_modifiable_params,
             )
-        except (RemoteError, DeserializationError):
-            if callback_modifiable_params.last_block_processed != NO_BLOCK_PROCESSED_VALUE:
-                with suppress(RemoteError):
-                    block = self.evm_inquirer.get_block_by_number(
-                        num=callback_modifiable_params.last_block_processed,
-                    )
-                    last_queried_ts = block['timestamp']
-        else:
-            last_queried_ts = to_ts
-        finally:
-            if last_queried_ts is not None:
-                with self.database.user_write() as write_cursor:
-                    for address in addresses:
-                        self.database.update_used_query_range(
-                            write_cursor=write_cursor,
-                            name=f'{BRIDGE_QUERIED_ADDRESS_PREFIX}{address}',
-                            start_ts=from_ts,
-                            end_ts=last_queried_ts,
-                        )
-            else:
-                log.debug(f'Gnosis bridge logs not queried. Not saving any query range for {address} in range {from_ts} to {to_ts}')  # noqa: E501
+        except (RemoteError, DeserializationError) as e:
+            log.error(f'Failed to query gnosis logs in full range from {from_ts} to {to_ts} for {addresses} due to {e}. Skipping for now')  # noqa: E501