+++ docs/changelog.rst
@@ -11,6 +11,7 @@ Changelog
 * :bug:`8892` rotki will now correctly fetch Starknet token prices before May 2024 from Cryptocompare, when the ticker changed from STARK to STRK.
 * :bug:`-` The airdrops directory should no longer appear under the user directory in certain circumstances.
 * :bug:`-` Fix an issue that caused the token detection to fail under some circumstances involving broken tokens.
+* :bug:`-` rotki won't try to query logs from slow nodes.
 
 * :release:`1.36.0 <2024-11-06>`
 * :bug:`-` The exported file that overrides the file with the same name should have the latest modified time.